# api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/bulk-sales/{id}": {
      "put": {
        "tags": [
          "campaign-controller"
        ],
        "operationId": "updateBulkSale",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBulkSaleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/marketplace/bulk-sales/{id}/offers": {
      "post": {
        "tags": [
          "marketplace-controller"
        ],
        "operationId": "createOffer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/chat/rooms": {
      "get": {
        "tags": [
          "chat-web-socket-controller"
        ],
        "operationId": "getMyRooms",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatRoomResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat-web-socket-controller"
        ],
        "operationId": "getOrCreateRoom",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatRoomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatRoomResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/rooms/{roomId}/messages": {
      "get": {
        "tags": [
          "chat-web-socket-controller"
        ],
        "operationId": "getMessages",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatMessageResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat-web-socket-controller"
        ],
        "operationId": "sendMessageRest",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendChatMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/marketplace/enterprises/me/offers": {
      "get": {
        "tags": [
          "marketplace-controller"
        ],
        "operationId": "getMyOffers",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OfferResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/marketplace/bulk-sales": {
      "get": {
        "tags": [
          "marketplace-controller"
        ],
        "operationId": "searchBulkSales",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BulkSaleResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/marketplace/bulk-sales/{id}": {
      "get": {
        "tags": [
          "marketplace-controller"
        ],
        "operationId": "getBulkSaleDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BulkSaleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/marketplace/bulk-sales/{id}/commitments": {
      "get": {
        "tags": [
          "marketplace-controller"
        ],
        "operationId": "getCommitmentsByBulkSale",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommitmentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/files/presigned-url": {
      "get": {
        "tags": [
          "file-upload-controller"
        ],
        "operationId": "getPresignedUrl",
        "parameters": [
          {
            "name": "filename",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contentType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/bulk-sales/{id}/offers": {
      "get": {
        "tags": [
          "campaign-controller"
        ],
        "operationId": "getOffers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OfferResponse"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SimpleResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "UpdateBulkSaleRequest": {
        "type": "object",
        "properties": {
          "expectedPrice": {
            "type": "number"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateOfferRequest": {
        "type": "object",
        "properties": {
          "offeredPrice": {
            "type": "number"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "CreateChatRoomRequest": {
        "type": "object",
        "properties": {
          "bulkSaleId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ChatRoomResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bulkSaleId": {
            "type": "string",
            "format": "uuid"
          },
          "productName": {
            "type": "string"
          },
          "cooperativeId": {
            "type": "string",
            "format": "uuid"
          },
          "cooperativeName": {
            "type": "string"
          },
          "cooperativeAvatarUrl": {
            "type": "string"
          },
          "enterpriseId": {
            "type": "string",
            "format": "uuid"
          },
          "enterpriseName": {
            "type": "string"
          },
          "enterpriseAvatarUrl": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SendChatMessageRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "senderType": {
            "type": "string"
          }
        }
      },
      "ChatMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "senderType": {
            "type": "string"
          },
          "senderName": {
            "type": "string"
          },
          "senderAvatarUrl": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "OfferResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "enterpriseName": {
            "type": "string"
          },
          "offeredPrice": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BulkSaleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "productName": {
            "type": "string"
          },
          "totalQuantity": {
            "type": "number"
          },
          "expectedPrice": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "NEGOTIATING",
              "SOLD",
              "CANCELLED"
            ]
          },
          "coopName": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CommitmentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "farmerName": {
            "type": "string"
          },
          "plotName": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PresignedUrlResponse": {
        "type": "object",
        "properties": {
          "presignedUrl": {
            "type": "string"
          },
          "publicUrl": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}